# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User settings",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User registration",
                "operationId": "authUser",
                "requestBody": {
                    "$ref": "#/components/requestBodies/RegisterRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true
                                        }
                                    },
                                    "bool": {
                                        "summary": "A boolean value.",
                                        "value": false
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "loginUser",
                "requestBody": {
                    "$ref": "#/components/requestBodies/LoginRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true
                                        }
                                    },
                                    "bool": {
                                        "summary": "A boolean value.",
                                        "value": false
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out.",
                "description": "Method revokes the current access token.",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true
                                        }
                                    },
                                    "bool": {
                                        "summary": "A boolean value.",
                                        "value": false
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Нет доступа"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user-settings/{userSetting}": {
            "post": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Update user setting",
                "description": "Update a user setting and send a confirmation code via the specified method.",
                "operationId": "bba3b8c14f2ebecf1bef5ad46a91d105",
                "parameters": [
                    {
                        "name": "userSetting",
                        "in": "path",
                        "description": "ID of the user setting",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateUserSettingRequest"
                },
                "responses": {
                    "200": {
                        "description": "Confirmation code sent"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "User setting not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{user}/settings/{setting}/confirm": {
            "post": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Confirm user setting",
                "description": "Confirm user setting by code and method",
                "operationId": "3887c46995503edf399ad7cc3aa1e534",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user setting",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "setting",
                        "in": "path",
                        "description": "setting ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/ConfirmUserSettingRequest"
                },
                "responses": {
                    "200": {
                        "description": "Setting confirmed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Setting confirmed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid confirmation code or method",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid confirmation code or method"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "requestBodies": {
            "ConfirmUserSettingRequest": {
                "description": "Confirm user settings.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "code",
                                "method",
                                "value"
                            ],
                            "properties": {
                                "code": {
                                    "description": "Confirmation code.",
                                    "type": "string",
                                    "example": "confCode"
                                },
                                "method": {
                                    "description": "Confirmation method.",
                                    "type": "string",
                                    "example": "confMethod"
                                },
                                "value": {
                                    "description": "new value.",
                                    "type": "string",
                                    "example": "value"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "LoginRequest": {
                "description": "Login users.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "email",
                                "password"
                            ],
                            "properties": {
                                "email": {
                                    "description": "The user email.",
                                    "type": "string",
                                    "format": "email",
                                    "example": "user@email.com"
                                },
                                "password": {
                                    "description": "The user password.",
                                    "type": "string",
                                    "example": "UserPassword"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RegisterRequest": {
                "description": "Register users.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "name",
                                "email",
                                "password"
                            ],
                            "properties": {
                                "name": {
                                    "description": "The user name.",
                                    "type": "string",
                                    "example": "UserName"
                                },
                                "email": {
                                    "description": "The user email.",
                                    "type": "string",
                                    "example": "UserEmail"
                                },
                                "password": {
                                    "description": "The user password.",
                                    "type": "string",
                                    "format": "password",
                                    "example": "UserPassword"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateUserSettingRequest": {
                "description": "Update user settings.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "method"
                            ],
                            "properties": {
                                "method": {
                                    "description": "The method.",
                                    "type": "string",
                                    "default": "email",
                                    "enum": [
                                        "email",
                                        "sms",
                                        "telegram"
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}